+++ pyogrio/tests/test_core.py
@@ -1,3 +1,5 @@
+import re
+import urllib.request
 from pathlib import Path
 
 import numpy as np
@@ -14,12 +16,14 @@ from pyogrio import (
     read_bounds,
     read_info,
     set_gdal_config_options,
+    vsi_curl_clear_cache,
     vsi_listtree,
     vsi_rmtree,
     vsi_unlink,
 )
-from pyogrio._compat import GDAL_GE_38
+from pyogrio._compat import GDAL_GE_38, GDAL_GE_311, GDAL_GE_312
 from pyogrio._env import GDALEnv
+from pyogrio.core import list_drivers_details
 from pyogrio.errors import DataLayerError, DataSourceError
 from pyogrio.raw import read, write
 from pyogrio.tests.conftest import (
@@ -41,6 +45,59 @@ try:
     import shapely
 except ImportError:
     pass
+
+
+@pytest.fixture()
+def driver_access_modes() -> dict[str, str]:
+    """Expected driver access modes adjusted for the active GDAL version."""
+    # Base expected driver access modes.
+    driver_access_modes: dict[str, str] = {
+        "AeronavFAA": "r",
+        "ARCGEN": "r",
+        "BNA": "rw",
+        "DXF": "rw",
+        "CSV": "raw",
+        "FileGDB": "raw",
+        "OpenFileGDB": "raw",
+        "ESRIJSON": "r",
+        "ESRI Shapefile": "raw",
+        "FlatGeobuf": "rw",  # Remark: append only if no spatial index
+        "GeoJSON": "raw",
+        "GeoJSONSeq": "raw",
+        "GPKG": "raw",
+        "GML": "rw",
+        "GMT": "rw",
+        "OGR_GMT": "rw",
+        "GPX": "rw",
+        "Idrisi": "r",
+        "MapInfo File": "raw",
+        "DGN": "rw",  # Remark: unclear if append is possible
+        # "Parquet": "rw",  # Newer versions of GDAL support append
+        "PCIDSK": "raw",
+        "PDS": "r",
+        "PGDUMP": "w",
+        "OGR_PDS": "r",
+        "S57": "rw",  # Remark: create supported according to GDAL docs
+        "SEGY": "r",
+        "SQLite": "raw",
+        "SUA": "r",
+        "TileDB": "raw",
+        "TopoJSON": "r",
+    }
+
+    # Update / append capability is only available from GDAL 3.11 onwards.
+    if not GDAL_GE_311:
+        driver_access_modes = {
+            name: mode.replace("a", "") for name, mode in driver_access_modes.items()
+        }
+
+    # GeoJSONSeq append capability is only available from GDAL 3.12 onwards.
+    if not GDAL_GE_312:
+        driver_access_modes["GeoJSONSeq"] = driver_access_modes["GeoJSONSeq"].replace(
+            "a", ""
+        )
+
+    return driver_access_modes
 
 
 def test_gdal_data():
@@ -134,17 +191,25 @@ def test_ogr_driver_supports_write(driver, expected):
     assert ogr_driver_supports_write(driver) == expected
 
 
-def test_list_drivers():
+def test_list_drivers(driver_access_modes):
     all_drivers = list_drivers()
 
-    # verify that the core drivers are present
-    for name in ("ESRI Shapefile", "GeoJSON", "GeoJSONSeq", "GPKG", "OpenFileGDB"):
-        assert name in all_drivers
-        expected_capability = "rw"
-        assert all_drivers[name] == expected_capability
+    # Verify that the core drivers are present + their capabilities match expectations.
+    for name, expected_capability in driver_access_modes.items():
+        if name not in all_drivers:
+            continue
 
+        assert all_drivers[name] == expected_capability, (
+            f"Error for {name}: {expected_capability=}, {all_drivers[name]=}"
+        )
+
+    # Check the filtering functionality of list_drivers
     drivers = list_drivers(read=True)
     expected = {k: v for k, v in all_drivers.items() if v.startswith("r")}
+    assert len(drivers) == len(expected)
+
+    drivers = list_drivers(append=True)
+    expected = {k: v for k, v in all_drivers.items() if "a" in v}
     assert len(drivers) == len(expected)
 
     drivers = list_drivers(write=True)
@@ -157,6 +222,59 @@ def test_list_drivers():
     }
     assert len(drivers) == len(expected)
 
+    drivers = list_drivers(read=True, write=True, append=True)
+    expected = {
+        k: v
+        for k, v in all_drivers.items()
+        if v.startswith("r") and v.endswith("w") and "a" in v
+    }
+    assert len(drivers) == len(expected)
+
+
+def test_list_drivers_details(driver_access_modes):
+    drivers = list_drivers_details()
+
+    # Verify detailed capabilities for all drivers that we track expected access for.
+    for name, expected_access_mode in driver_access_modes.items():
+        if name not in drivers:
+            continue
+
+        try:
+            assert drivers[name]["long_name"] is not None
+
+            assert drivers[name]["read"] is ("r" in expected_access_mode)
+            if GDAL_GE_311:
+                assert drivers[name]["append"] is ("a" in expected_access_mode)
+            else:
+                # GDAL < 3.11 does not support reporting update/append capability
+                assert drivers[name]["append"] is None
+            assert drivers[name]["write"] is ("w" in expected_access_mode)
+            assert drivers[name]["supports_vsi"]
+            assert drivers[name]["help_topic_url"] is None or isinstance(
+                drivers[name]["help_topic_url"], str
+            )
+            assert drivers[name]["extensions"] is None or isinstance(
+                drivers[name]["extensions"], list
+            )
+            if drivers[name]["extensions"] is not None:
+                assert all(ext.startswith(".") for ext in drivers[name]["extensions"])
+        except AssertionError as e:
+            raise AssertionError(f"Error for driver {name}: {e}") from e
+
+
+@pytest.mark.network
+def test_list_drivers_details_help_topic_url():
+    """Check if the help_topic_url for the GeoJSON driver is valid and reachable."""
+    drivers = list_drivers_details()["GeoJSON"]
+
+    request = urllib.request.Request(
+        drivers["help_topic_url"], method="HEAD", headers={"User-Agent": "pyogrio-test"}
+    )
+    ret = urllib.request.urlopen(request)
+    assert ret.status == 200, (
+        f"Help topic URL {drivers['help_topic_url']} is not reachable"
+    )
+
 
 def test_list_layers(
     naturalearth_lowres,
@@ -168,14 +286,16 @@ def test_list_layers(
     multisurface_file,
     no_geometry_file,
 ):
+    # SHP
+    geom_type = "Polygon" if __gdal_version__ < (3, 14) else "MultiPolygon"
     assert array_equal(
-        list_layers(naturalearth_lowres), [["naturalearth_lowres", "Polygon"]]
+        list_layers(naturalearth_lowres), [["naturalearth_lowres", geom_type]]
+    )
+    assert array_equal(
+        list_layers(naturalearth_lowres_vsi[1]), [["naturalearth_lowres", geom_type]]
     )
 
-    assert array_equal(
-        list_layers(naturalearth_lowres_vsi[1]), [["naturalearth_lowres", "Polygon"]]
-    )
-
+    # in-memory GPKG
     assert array_equal(
         list_layers(naturalearth_lowres_vsimem),
         [["naturalearth_lowres", "MultiPolygon"]],
@@ -184,7 +304,7 @@ def test_list_layers(
     # Measured 3D is downgraded to plain 3D during read
     # Make sure this warning is raised
     with pytest.warns(
-        UserWarning, match=r"Measured \(M\) geometry types are not supported"
+        UserWarning, match=re.escape("Measured (M) geometry types are not supported")
     ):
         assert array_equal(list_layers(line_zm_file), [["line_zm", "LineString Z"]])
 
@@ -445,7 +565,11 @@ def test_read_info(naturalearth_lowres):
         # geometry_name == "" for formats where geometry column name cannot be
         # customized
         assert meta["geometry_name"] == ""
-        assert meta["geometry_type"] == "Polygon"
+        assert (
+            meta["geometry_type"] == "Polygon"
+            if __gdal_version__ < (3, 14)
+            else "MultiPolygon"
+        )
         assert meta["driver"] == "ESRI Shapefile"
         assert meta["capabilities"]["fast_set_next_by_index"] is True
     else:
@@ -700,3 +824,13 @@ def test_vsimem_unlink_error(naturalearth_lowres_vsimem):
 
     with pytest.raises(FileNotFoundError, match="Path does not exist"):
         vsi_unlink("/vsimem/non-existent.gpkg")
+
+
+def test_vsi_curl_clear_cache_empty_default():
+    # Validate call doesn't raise any error when no prefix is used
+    vsi_curl_clear_cache()
+
+
+def test_vsi_curl_clear_cache():
+    # Validate call doesn't raise any error when prefix/path is used
+    vsi_curl_clear_cache("http://example.com/prefix")