+++ CHANGES.md
@@ -1,18 +1,43 @@
 # CHANGELOG
 
-## 0.12.2 (????-??-??)
+
+## 0.14.0 (????-??-??)
+
+
+### Packaging
+
+-   Minimum required Python version is now 3.11 (#655).
+
+## 0.13.0 (2026-06-26)
+
+### Improvements
+
+-   Support reading Time type columns by default (was already supported with `use_arrow=False`) (#617).
+-   Add `list_drivers_details()` function to list the available drivers with more
+    detailed properties like the long driver names, whether read/append/write is
+    supported by the driver, the help topic URL, etc (#656).
+-   Add whether append is supported in `list_drivers` (#559)
+-   Add `vsi_curl_clear_cache()` function to allow users to clear the local GDAL vsi cache
+    associated with /vsicurl/ (and related file systems) (#605).
+-   Improve performance of `read_dataframe` (especially if a filter is used)
+    and `write_dataframe` without Arrow (#577, #674).
+-   Unlock the GIL during GDAL functions that can take significant time to improve
+    performance when multithreading (#572).
 
 ### Bug fixes
 
+-   Fix error in `write_dataframe` with `use_arrow=False` when writing an object-type
+    column with datetimes without any time zone offsets or with a mix of offsets and
+    no offsets (#634).
 -   Fix writing non-string object columns with arrow (#630).
 -   Fix writing empty string category columns with arrow fails (#621).
--   Fix Time type columns being skipped with `use_arrow=False` (#617).
 -   Fix overwriting a corrupt fileGDB directory (#600).
 -   Fix attribute data being incorrectly written with KML driver (#650).
 
 ### Packaging
 
--   Minimum required Python version is now 3.11 (#xxx).
+-   The GDAL library included in the wheels is upgraded from 3.11.4 to 3.12.4 (#658).
+-   The wheels for Python 3.11+ are now distributed on PyPI as a single ABI3 wheel per platform (#660).
 
 ## 0.12.1 (2025-11-28)